+++ src/components/DataDisplay/FileViewer/FileViewer.tsx
@@ -1,7 +1,9 @@
-import { Box, SxProps, Theme } from "@mui/material";
-import { PropsWithChildren, useCallback, useEffect, useRef, useState } from "react";
-import notFoundImage from "@/assets/img/not-found-img.svg";
+import { Box, SxProps, Theme, Typography } from "@mui/material";
+import { PropsWithChildren, useState } from "react";
+import notFoundImage from "@/assets/img/not-found-img.jpg";
+import sheetsImage from "@/assets/img/sheets.png";
 import Lightbox from "@/components/Feedback/Lightbox";
+import isDocumentType from "@/utils/isDocumentType";
 
 interface FileViewerPros extends PropsWithChildren {
   src?: string | null;
@@ -29,6 +31,14 @@ const styles = {
     flexShrink: 0,
     overflow: "hidden",
     position: "relative",
+    userSelect: "none",
+  },
+  extension: {
+    backgroundColor: ({ palette }: Theme) => palette.common.white,
+    bottom: "10%",
+    paddingX: 0.8,
+    position: "absolute",
+    right: 0,
   },
   thumb: {
     "&::-webkit-scrollbar": {
@@ -39,7 +49,6 @@ const styles = {
     cursor: "pointer",
     display: "block",
     msOverflowStyle: "none",
-    objectFit: "cover",
     overflow: "hidden",
     pointerEvents: "none !important",
     scrollbarWidth: "none",
@@ -66,8 +75,6 @@ const FileViewer = ({
   srcThumb,
   fileName,
   children,
-  width,
-  height,
   sx,
   widthLightbox,
   heightLightbox,
@@ -78,21 +85,30 @@ const FileViewer = ({
   onClose,
   onClickThumb,
   variant,
+  height = 152,
+  width = 220,
 }: FileViewerPros) => {
   const [internalOpen, setInternalOpen] = useState(false);
   const [isError, setIsError] = useState(false);
+  const [isLoading, setIsLoading] = useState(true);
   const lowercaseSrc = src?.toLowerCase();
   const isImage = !lowercaseSrc?.endsWith(".pdf") && !lowercaseSrc?.startsWith("blob:") && !lowercaseSrc?.endsWith(".eml");
   const isPdf = lowercaseSrc?.endsWith(".pdf");
-  const iframeRef = useRef<HTMLObjectElement>(null);
-  const data = isError ? notFoundImage : srcThumb || src;
   const opacity = disableLightbox ? 1 : 0.8;
+  const isThumbnailReady = disableThumb ? true : !isLoading;
+  const isDocument = isDocumentType(src) || isDocumentType(srcThumb);
+  const shouldDisplayLightbox = (isImage || isPdf) && !isDocument && isThumbnailReady && !isError && !disableLightbox && src;
+  const extension = src?.split(".").pop()?.toLowerCase();
+
+  const getSrcThumb = () => {
+    if (isDocument) {
+      return sheetsImage;
+    }
+
+    return isError ? notFoundImage : srcThumb || src || undefined;
+  };
 
   const toggleOpen = () => {
-    // prevent mixed content to be displayed.
-    if (!isImage && !isPdf) {
-      return;
-    }
     setInternalOpen((prevState) => !prevState);
   };
 
@@ -101,21 +117,18 @@ const FileViewer = ({
     onClose?.();
   };
 
-  const errorHandle = useCallback(() => {
+  const handleError = () => {
     setIsError(true);
-  }, []);
+  };
 
-  /**
-   * Handle error
-   */
-  useEffect(() => {
-    const object = iframeRef.current;
-    object?.addEventListener("error", errorHandle);
+  const handleLoad = () => {
+    setIsLoading(false);
+  };
 
-    return () => {
-      object?.removeEventListener("error", errorHandle);
-    };
-  }, [errorHandle, iframeRef]);
+  const handleClick = () => {
+    onClickThumb?.();
+    toggleOpen();
+  };
 
   return (
     <>
@@ -124,10 +137,7 @@ const FileViewer = ({
           data-test="fileViewer"
           width={width}
           height={height}
-          onClick={() => {
-            onClickThumb?.();
-            toggleOpen();
-          }}
+          onClick={handleClick}
           sx={{
             ...styles.container,
             ":hover": { opacity },
@@ -139,19 +149,30 @@ const FileViewer = ({
           }}
         >
           <Box
-            component={isImage ? "img" : "iframe"}
-            key={data}
             overflow="hidden"
-            height={isImage ? "100%" : "auto"}
             width="100%"
-            src={data || undefined}
-            ref={iframeRef}
-            sx={styles.thumb}
+            component={isImage ? "img" : "iframe"}
+            height={isImage ? "100%" : "auto"}
+            key={getSrcThumb()}
+            src={getSrcThumb()}
+            onError={handleError}
+            onLoad={handleLoad}
+            sx={{
+              ...styles.thumb,
+              objectFit: isDocument ? "contain" : "cover",
+              padding: isDocument ? "15%" : 0,
+            }}
           />
+          {isDocument && (
+            <Typography sx={styles.extension} variant="body3" color="black">
+              {extension}
+            </Typography>
+          )}
         </Box>
       )}
 
-      {!disableLightbox && src && (
+      {/* Lightbox */}
+      {shouldDisplayLightbox && (
         <Lightbox open={open !== undefined && (isPdf || isImage) ? open : internalOpen} onClose={close} src={src} title={fileName}>
           <Box
             component={isPdf ? "iframe" : "img"}